ysib/requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
#channels==2.4.0
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect2==0.4.1
daphne==2.4.1
defusedxml==0.7.1
Django==2.2.22
#django-channels==0.7.0
django-crispy-forms==1.11.2
django-qr-code==2.1.0
django-user-agents==0.4.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
oauthlib==3.1.0
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
Pyphen==0.10.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
segno==1.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
WeasyPrint==52.5
webencodings==0.5.1
zope.interface==5.4.0
